Refactor user profile component to use `Signal`s
This is a (somewhat ugly) start on getting `UserProfileComponent` to use `Signal`s instead of `Observable`s, but it does work.

The error handling is a bit icky because we end up returning either a `User` or an unnamed error type. At a minimum, we should create a named error type to simplify all the checking and conversion.

I have a feeling that we might also be able to have `getUserById` return a compound type that contains either a `User` or an error _and knows which it has_. This might let us avoid a lot of this annoying testing/conversion logic, or at least put it all in one sensible place.

The E2E tests pass, but the unit (Karma) tests do not.