Bump the add user timeout to 5 full minutes
We keep having issues (which I assume are timeouts) in our Cypress tests for adding a user when they run in GitHub Actions. This takes the rather extreme approach of allowing up to _five_ full minutes on those assertions in the hopes that this stops that flakiness in our tests.