Update from Angular 18 to 19
This process was spelled out in Angular documentation: https://angular.dev/update-guide?v=18.0-19.0&l=3

Most of the changes worked without any trouble when I entered the correct set of things to update in a given step (which was a little different from the checklist). One component test was failing and required some extra attention. I will put that in the next commit since I'm not totally sure it's strictly related to this update (it was passing before the update, and it used a deprecated module that when I removed it seemed to solve the problem, but I did also update a stub).

Most of the changes that you will see here are due to the fact that standalone is the default for components now, so that line of code was automatically removed in several places in the `ng update` process.