Update `userServiceStub` and deprecated use of `RouterTestingModule`
I'm not 100% sure why the move from Angular 18 to 19 caused this test to fail. I initially thought it might be due to the deprecated use of `RouterTestingModule`. In the process of updating the deprecated use of `RouterTestingModule`, I noticed that I was getting the error that my UserService had no method `filterUsers`. I noticed, too, that the actual UserService has such a method, so I initially tried adding to the mocked user service in the testing folder. When that error persisted, I noticed that the test about the misbehaving user list used a stubbed user service, so I updated the userServiceStub minimally to make it return [] when it's asked to to the filterUsers method.

I'm not particularly confident I did that correctly, but it was the minimal code to make the test pass, so I thought that was a good sign.

This removes one instance of RouterTestingModule, which is listed as an issue https://github.com/UMM-CSci-3601/3601-iteration-template/issues/1581

After making this work, I also removed the rest of those in a way that I think is safe. I'll make that on a separate branch since it seems like its own thing.